Fixed the error with the inaccurate targetting. Turns out, the error had to do with a rounding issue. Changed the xPosition and yPosition to floats to remedy the problem.

+++ BulletHell/Bullet.pde
@@ -1,7 +1,7 @@
 int countdown;
 
 public class Bullet {
-  int xpos, ypos;
+  float xpos, ypos;
   float bulletXSpeed, bulletYSpeed;
 
   public Bullet(int x, int y) {
